requirements.txt
aws-cdk-lib==2.151.0
constructs>=10.0.0,<11.0.0
streamlit
pypdf
langchain
faiss-cpu
boto3
langchain_community
retrying
opensearch-py